+++ src/idvc/dvc_interface.py
@@ -14,6 +14,7 @@
 #   Author: Edoardo Pasca (UKRI-STFC)
 
 import os
+from openpyxl import load_workbook
 import sys
 import PySide2
 from PySide2 import QtCore, QtGui, QtWidgets
@@ -99,6 +100,7 @@ import logging
 
 from idvc.utils.AutomaticRegistration import AutomaticRegistration
 
+allowed_point_cloud_file_formats = ('.txt','.csv','.xlsx')
 class MainWindow(QMainWindow):
     def __init__(self):
         QMainWindow.__init__(self)
@@ -2895,7 +2897,8 @@ Each line in the tab delimited file contains an integer point label followed by 
 2   300   750.2  209\n\
 etc.\n\
 Non-integer voxel locations are admitted, with reference volume interpolation used as needed.\n\
-The first point is significant, as it is used as a global starting point and reference for the rigid_trans variable.")
+The first point is significant, as it is used as a global starting point and reference for the rigid_trans variable.\n\
+File format allowed: 'txt', 'csv, 'xlxs'.")
         pc['roi_browse'].clicked.connect(self.select_pointcloud)
         self.graphWidgetFL.setWidget(widgetno, QFormLayout.FieldRole, pc['roi_browse'])
         widgetno += 1
@@ -3072,6 +3075,9 @@ The first point is significant, as it is used as a global starting point and ref
                 self.overlapXValueEntry.setEnabled(False)
 
     def select_pointcloud(self): #, label):
+        """Opens a dialog to select the pointcloud from a file. 
+        Runs the creation or loading of the point cloud in a worker.
+        Sets the attribut `self.pointcloud_is` to 'loaded'."""
         dialogue = QFileDialog()
         self.roi = None
         self.roi = dialogue.getOpenFileName(self,"Select a roi")[0]
@@ -3089,6 +3095,8 @@ The first point is significant, as it is used as a global starting point and ref
             self.pointcloud_is = 'loaded'
 
     def PointCloudWorker(self, type, filename = None, disp_file = None, vector_dim = None):
+        """Runs the worker to create or load the point cloud.
+        If the format of the point-cloud file is not in the allowed list it displays an error dialog."""
         if type == "create":
             #if not self.pointCloudCreated:
             self.clearPointCloud()
@@ -3100,8 +3108,15 @@ The first point is significant, as it is used as a global starting point and ref
             self.pointcloud_worker.signals.result.connect(self.DisplayLoadedPointCloud)
         elif type == "load pointcloud file":
             self.clearPointCloud()
-            self.pointcloud_worker = Worker(self.loadPointCloud, self.roi)
-            self.pointcloud_worker.signals.result.connect(self.DisplayLoadedPointCloud)
+            if self.roi.endswith(allowed_point_cloud_file_formats):
+                self.pointcloud_worker = Worker(self.loadPointCloud, self.roi)
+                self.pointcloud_worker.signals.result.connect(self.DisplayLoadedPointCloud)
+            else:
+                error_title = "FILE FORMAT ERROR"
+                error_text = f"Error reading the point-cloud file {self.roi}. Allowed formats are {allowed_point_cloud_file_formats}."
+                self.displayFileErrorDialog(message=error_text, title=error_title)
+                return
+                
         elif type == "create without loading":
             #if not self.pointCloudCreated:
             self.clearPointCloud()
@@ -3414,6 +3429,9 @@ The first point is significant, as it is used as a global starting point and ref
         return True
             
     def loadPointCloud(self, *args, **kwargs):
+        """Loads a pointcloud from file. 
+        File formats allowed are 'txt', 'csv', 'xlxs'. 
+        """
         time.sleep(0.1) #required so that progress window displays
         pointcloud_file = os.path.abspath(args[0])
         progress_callback = kwargs.get('progress_callback', None)
@@ -3421,8 +3439,14 @@ The first point is significant, as it is used as a global starting point and ref
         #self.clearPointCloud() #need to clear current pointcloud before we load next one TODO: move outside thread
         progress_callback.emit(30)
         self.roi = pointcloud_file
-
-        points = np.loadtxt(self.roi)
+        if pointcloud_file.endswith('.txt'):
+            points = np.loadtxt(pointcloud_file)
+        elif pointcloud_file.endswith('.csv'):
+            points = np.genfromtxt(pointcloud_file, delimiter=',')
+        elif pointcloud_file.endswith('.xlsx'):
+            workbook = load_workbook(pointcloud_file, read_only=True)
+            sheet = workbook.active
+            points = np.array(list(sheet.values))
         # except ValueError as ve:
         #     print(ve)
         #     return